// consequences/src/lib/sbs/dsl/compiler/types.ts
export const TokenType = {
  KEYWORD: 'KEYWORD',
  IDENT: 'IDENT',
  NUMBER: 'NUMBER',
  STRING: 'STRING',
  OP: 'OP',
  PUNC: 'PUNC',
  ARROW: 'ARROW',
  PIPE: 'PIPE',
  EOF: 'EOF',
} as const;

export type TokenType = typeof TokenType[keyof typeof TokenType];

export interface Token {
  type: TokenType;
  value: string | number | null;
  line: number;
  col: number;
}